requirements.txt
alembic==1.7.7
APScheduler==3.9.1
Authlib==1.0.1
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
dominate==2.6.0
elastic-transport==8.1.2
elasticsearch==8.2.3
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-APScheduler==1.12.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.0.1
Jinja2==3.1.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
pdf2image==1.16.0
Pillow==9.1.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
svgwrite==1.4.2
Tree==0.2.4
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.8.0
